refactor: simplify logger
